% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feel.R
\docType{data}
\name{sentiments_polarity}
\alias{sentiments_polarity}
\title{sentiments_polarity}
\format{data.frame}
\source{
Amine Abdaoui, Jérôme Azé, Sandra Bringay et Pascal Poncelet. FEEL: French Expanded Emotion Lexicon. Language Resources and Evaluation, LRE 2016, pp 1-23.
}
\usage{
sentiments_polarity
}
\description{
The sentiment polarity dataset
}
\keyword{data}
