% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feel.R
\docType{data}
\name{sentiments_score}
\alias{sentiments_score}
\title{sentiments_score}
\format{data.frame}
\source{
Amine Abdaoui, Jérôme Azé, Sandra Bringay et Pascal Poncelet. FEEL: French Expanded Emotion Lexicon. Language Resources and Evaluation, LRE 2016, pp 1-23.
}
\usage{
sentiments_score
}
\description{
The sentiment score dataset
}
\keyword{data}
